requirements.txt
# Python dependencies for the bonk.io RL environment
pyzmq>=25.0.0
gymnasium>=0.29.0
numpy>=1.24.0
